# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DND AI Roleplay",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/campaigns/{campaign_id}/dm-drafts": {
      "post": {
        "tags": [
          "dm-drafts"
        ],
        "summary": "Create Draft",
        "operationId": "create_draft_api_v1_campaigns__campaign_id__dm_drafts_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DmDraftCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DmDraftView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "dm-drafts"
        ],
        "summary": "List Drafts",
        "operationId": "list_drafts_api_v1_campaigns__campaign_id__dm_drafts_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DmDraftView"
                  },
                  "title": "Response List Drafts Api V1 Campaigns  Campaign Id  Dm Drafts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/dm-drafts:generate-opening": {
      "post": {
        "tags": [
          "dm-drafts"
        ],
        "summary": "Generate Opening",
        "operationId": "generate_opening_api_v1_campaigns__campaign_id__dm_drafts_generate_opening_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DmDraftView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dm-drafts/{draft_id}": {
      "patch": {
        "tags": [
          "dm-drafts"
        ],
        "summary": "Update Draft",
        "operationId": "update_draft_api_v1_dm_drafts__draft_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DmDraftUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DmDraftView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dm-drafts/{draft_id}:discard": {
      "post": {
        "tags": [
          "dm-drafts"
        ],
        "summary": "Discard Draft",
        "operationId": "discard_draft_api_v1_dm_drafts__draft_id__discard_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DmDraftView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dm-drafts/{draft_id}:publish": {
      "post": {
        "tags": [
          "dm-drafts"
        ],
        "summary": "Publish Draft",
        "operationId": "publish_draft_api_v1_dm_drafts__draft_id__publish_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DmDraftPublishResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{campaign_id}/messages": {
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_api_v1_campaigns__campaign_id__messages_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageView"
                  },
                  "title": "Response List Messages Api V1 Campaigns  Campaign Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Send Dm Message",
        "operationId": "send_dm_message_api_v1_campaigns__campaign_id__messages_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DmMessageCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DmMessageCommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/messages:ooc": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Correct Message With Ooc",
        "operationId": "correct_message_with_ooc_api_v1_campaigns__campaign_id__messages_ooc_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OocCorrectionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/runtime:stop": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Stop Runtime",
        "operationId": "stop_runtime_api_v1_campaigns__campaign_id__runtime_stop_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/runtime:retry": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "Retry Runtime",
        "operationId": "retry_runtime_api_v1_campaigns__campaign_id__runtime_retry_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DmDraftCreate": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 6000,
            "minLength": 1,
            "title": "Prompt"
          },
          "sourceSkillCheckId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceskillcheckid"
          },
          "assistMode": {
            "type": "string",
            "pattern": "^(POLISH|EXPAND|REWRITE)$",
            "title": "Assistmode",
            "default": "POLISH"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "DmDraftCreate",
        "description": "`prompt` is the DM's own rough reply or a one-line intent."
      },
      "DmDraftPublishResult": {
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/DmDraftView"
          },
          "message": {
            "$ref": "#/components/schemas/MessageView"
          }
        },
        "type": "object",
        "required": [
          "draft",
          "message"
        ],
        "title": "DmDraftPublishResult"
      },
      "DmDraftStatus": {
        "type": "string",
        "enum": [
          "GENERATING",
          "READY",
          "STALE",
          "SENT",
          "DRAFT",
          "PUBLISHED",
          "DISCARDED",
          "FAILED"
        ],
        "title": "DmDraftStatus"
      },
      "DmDraftTriggerType": {
        "type": "string",
        "enum": [
          "OPENING",
          "CHARACTER_REPLY",
          "SKILL_CHECK_RESULT",
          "MANUAL_ASSIST"
        ],
        "title": "DmDraftTriggerType"
      },
      "DmDraftUpdate": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          },
          "content": {
            "type": "string",
            "maxLength": 1200,
            "minLength": 1,
            "title": "Content"
          },
          "audience": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ],
            "title": "Audience",
            "default": "PUBLIC"
          },
          "recipientCharacterIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Recipientcharacterids"
          }
        },
        "type": "object",
        "required": [
          "revision",
          "content"
        ],
        "title": "DmDraftUpdate"
      },
      "DmDraftView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "campaignId": {
            "type": "string",
            "title": "Campaignid"
          },
          "sessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sessionid"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "audience": {
            "$ref": "#/components/schemas/MessageAudience"
          },
          "recipientCharacterIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recipientcharacterids"
          },
          "improvisedNotes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Improvisednotes"
          },
          "status": {
            "$ref": "#/components/schemas/DmDraftStatus"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "sourceSkillCheckId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourceskillcheckid"
          },
          "sourceMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcemessageid"
          },
          "triggerType": {
            "$ref": "#/components/schemas/DmDraftTriggerType"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "campaignId",
          "sessionId",
          "content",
          "audience",
          "recipientCharacterIds",
          "improvisedNotes",
          "status",
          "revision",
          "sourceSkillCheckId",
          "sourceMessageId",
          "triggerType",
          "createdAt",
          "updatedAt"
        ],
        "title": "DmDraftView"
      },
      "DmMessageCommandResult": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/MessageView"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeState"
          }
        },
        "type": "object",
        "required": [
          "message",
          "runtime"
        ],
        "title": "DmMessageCommandResult"
      },
      "DmMessageCreate": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 1,
            "title": "Content"
          },
          "audience": {
            "$ref": "#/components/schemas/MessageAudience",
            "default": "PUBLIC"
          },
          "recipientCharacterIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Recipientcharacterids"
          },
          "addressedCharacterIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Addressedcharacterids"
          },
          "clientRequestId": {
            "type": "string",
            "format": "uuid",
            "title": "Clientrequestid"
          }
        },
        "type": "object",
        "required": [
          "content",
          "clientRequestId"
        ],
        "title": "DmMessageCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageAudience": {
        "type": "string",
        "enum": [
          "PUBLIC",
          "PRIVATE",
          "DM_ONLY"
        ],
        "title": "MessageAudience"
      },
      "MessageKind": {
        "type": "string",
        "enum": [
          "IN_GAME",
          "OOC",
          "SYSTEM"
        ],
        "title": "MessageKind"
      },
...
      "MessageSenderType": {
        "type": "string",
        "enum": [
          "DM",
          "CHARACTER",
          "SYSTEM"
        ],
        "title": "MessageSenderType"
      },
      "MessageView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "campaignId": {
            "type": "string",
            "title": "Campaignid"
          },
          "sequenceNo": {
            "type": "integer",
            "title": "Sequenceno"
          },
          "senderType": {
            "$ref": "#/components/schemas/MessageSenderType"
          },
          "senderCharacterId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sendercharacterid"
          },
          "senderName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sendername"
          },
          "kind": {
            "$ref": "#/components/schemas/MessageKind"
          },
          "audience": {
            "$ref": "#/components/schemas/MessageAudience"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "addressedCharacterIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Addressedcharacterids"
          },
          "isOocCorrected": {
            "type": "boolean",
            "title": "Isooccorrected"
          },
          "oocCorrectionNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ooccorrectionnote"
          },
          "recipients": {
            "items": {
              "$ref": "#/components/schemas/MessageRecipientView"
            },
            "type": "array",
            "title": "Recipients"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "campaignId",
          "sequenceNo",
          "senderType",
          "senderCharacterId",
          "senderName",
          "kind",
          "audience",
          "content",
          "addressedCharacterIds",
          "isOocCorrected",
          "oocCorrectionNote",
          "recipients",
          "createdAt"
        ],
        "title": "MessageView"
      },
...
      "OocCorrectionCreate": {
        "properties": {
          "targetMessageId": {
            "type": "string",
            "title": "Targetmessageid"
          },
          "correction": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 1,
            "title": "Correction"
          },
          "replacementContent": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3000
              },
              {
                "type": "null"
              }
            ],
            "title": "Replacementcontent"
          },
          "clientRequestId": {
            "type": "string",
            "format": "uuid",
            "title": "Clientrequestid"
          }
        },
        "type": "object",
        "required": [
          "targetMessageId",
          "correction",
          "clientRequestId"
        ],
        "title": "OocCorrectionCreate"
      },
...
      "RuntimeState": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/RuntimeStatus"
          },
          "generation": {
            "type": "integer",
            "title": "Generation"
          },
          "activeAgentRunId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activeagentrunid"
          },
          "waitingRequest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waitingrequest"
          },
          "lastErrorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasterrorcode"
          },
          "lastErrorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasterrormessage"
          },
          "consecutiveAiMessages": {
            "type": "integer",
            "title": "Consecutiveaimessages"
          }
        },
        "type": "object",
        "required": [
          "status",
          "generation",
          "activeAgentRunId",
          "waitingRequest",
          "lastErrorCode",
          "lastErrorMessage",
          "consecutiveAiMessages"
        ],
        "title": "RuntimeState"
      },
      "RuntimeStatus": {
        "type": "string",
        "enum": [
          "IDLE",
          "AGENTS_EVALUATING",
          "VALIDATING_MESSAGE",
          "WAITING_FOR_DM",
          "ERROR",
          "ENDED"
        ],
        "title": "RuntimeStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}